dont use posix_spawn in ninja for ios 1

+++ pkgs/ninja/fetch.sh
@@ -1,8 +1,12 @@
 #!/bin/sh
 rm -rf pkg src
-printf "Downloading source...\n"
-curl -L -# -o src.tar.gz https://github.com/ninja-build/ninja/archive/refs/tags/v1.13.1.tar.gz
+ver='1.13.1'
+[ -z "$_DLCACHE" ] && _DLCACHE=/tmp
+if [ ! -f "$_DLCACHE/ninja-$ver.tar.gz" ] ||
+    [ "$(sha256sum "$_DLCACHE/ninja-$ver.tar.gz" | awk '{print $1}')" != "f0055ad0369bf2e372955ba55128d000cfcc21777057806015b45e4accbebf23" ]; then
+    printf "Downloading source...\n"
+    curl -L -# -o "$_DLCACHE/ninja-$ver.tar.gz" "https://github.com/ninja-build/ninja/archive/refs/tags/v$ver.tar.gz" || exit 1
+fi
 printf "Unpacking source...\n"
-tar -xf src.tar.gz
-rm src.tar.gz
+tar -xf "$_DLCACHE/ninja-$ver.tar.gz"
 mv ninja-* src
